Create daily.py
This fails at line 47 with TypeError: "unbound method checkAction() must be...". It would have been helpful to see the rest of the truncated message. What must it be that it isn't. I also have no idea why checkAction() is unbound.  What does that even mean? How do I debug this?